this is my first commit. I am adding a discussion tag to let myself know that I intend to discuss this with them
this is my description that I like to use. It comes in handy when I need
to elaborate on the details of this commit. It also helps remind myself
in the future where this important change was and the reasoning behind
it. Sometimes we work on different projects and context switcing causes
goldfish brain. But have no worry, for we have good, well-written commit
descriptions.